requirements.txt
spotipy
Pillow
openai